Break out widget creation to its own module
